chore: Better implementation of styles
