+++ packages/article-magazine-standard/src/styles/native.js
@@ -1,0 +1,11 @@
+import sharedStyles from "./shared";
+
+const nativeStyles = {
+  ...sharedStyles,
+  standFirst: {
+    ...sharedStyles.standFirst,
+    lineHeight: 25
+  }
+};
+
+export default nativeStyles;
